requirements.txt
flask
torch
numpy
Pillow
matplotlib